Data lakes (HydroLAKES mask, water still emergent), measured floodplain
fractions replacing all three fertility painters, seasonal river flow with
freeze closure. Review: snow store gains a spin-up year (December snow now
feeds the spring melt crest); the unledgered glacier runoff term is deleted
— every river-seasons anchor passes without it. Conflicts: QUESTIONS
renumbered (their entry → #26), gradient sentinel keeps
TRAVEL_RIVER_GRADIENT_UNMEASURED. Full suite green: lint, smoke, gate
(lakes 10/10, floodplain 4/4, seasons 6/6, rivers 13/13, zero manifest
churn), oracle ok on all three arms.

+++ v2/src/sim/travel/battery.ts
@@ -87,7 +87,7 @@ function fixtureSubstrate(grid: GridPreset): Substrate {
       originFx: 0,
       originFy: 0,
     },
-    floodplain: new Uint8Array(N),
+    floodplain: new Float32Array(N),
     biome: new Uint8Array(N),
     soil: new Float32Array(N),
     fertility: new Float32Array(N),
